f fmt, test

+++ autopush-common/src/lib.rs
@@ -51,12 +51,12 @@ pub mod util;
 
 /// The maximum TTL for notifications (30 days).
 /// In most use cases, converted to seconds through .num_seconds().
-pub const MAX_NOTIFICATION_TTL_SECS: u32 = 30 * 86400;
+pub const MAX_NOTIFICATION_TTL_SECS: u64 = 30 * 86400;
 
 /// FCM has a max TTL of 4 weeks (28 days).
 /// In most use cases, converted to seconds through .num_seconds().
-pub const MAX_FCM_NOTIFICATION_TTL_SECS: u32 = 28 * 86400;
+pub const MAX_FCM_NOTIFICATION_TTL_SECS: u64 = 28 * 86400;
 
 /// The maximum TTL for router records (60 days).
 /// In most use cases, converted to seconds through .num_seconds().
-pub const MAX_ROUTER_TTL_SECS: u32 = 60 * 86400;
+pub const MAX_ROUTER_TTL_SECS: u64 = 60 * 86400;
